feat: extract command into its own package
